break DeviceFilter into smaller componnets
